Moves HealthCheckIQ to separate IQ provider.
