+++ src/net/java/sip/communicator/impl/protocol/jabber/extensions/colibri/ColibriIQProvider.java
@@ -900,32 +900,6 @@ public class ColibriIQProvider
                 }
             }
         }
-        else if (HealthCheckIQ.ELEMENT_NAME.equals(parser.getName())
-            && HealthCheckIQ.NAMESPACE.equals(namespace))
-        {
-            String rootElement = parser.getName();
-
-            iq = new HealthCheckIQ();
-
-            boolean done = false;
-
-            while (!done)
-            {
-                switch (parser.next())
-                {
-                    case XmlPullParser.END_TAG:
-                    {
-                        String name = parser.getName();
-
-                        if (rootElement.equals(name))
-                        {
-                            done = true;
-                        }
-                        break;
-                    }
-                }
-            }
-        }
         else
             iq = null;
